WeatherType Refactor
